breakout customized bridge.
